don't throw 400 if there were no tags specified in the body

+++ routes/api/product-routes.js
@@ -99,33 +99,39 @@ router.put('/:id', (req, res) => {
       return ProductTag.findAll({ where: { id: req.params.id } });
     })
     .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            id: req.params.id,
-            tag_id,
-          };
-        });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
+      let productTagsToRemove = [];
+      let newProductTags = [];
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
-    })
-    .then((updatedProductTags) => res.json(updatedProductTags))
-    .catch((err) => {
-      console.log(err);
-      res.status(400).json(err);
-    });
+      // define the productTagsToRemove and newProductTags if there were tags provided
+      if (req.body.tagIds) {
+        // get list of current tag_ids
+        const productTagIds = productTags.map(({ tag_id }) => tag_id);
+        // create filtered list of new tag_ids
+        newProductTags = req.body.tagIds
+          .filter((tag_id) => !productTagIds.includes(tag_id))
+          .map((tag_id) => {
+            return {
+              id: req.params.id,
+              tag_id,
+            };
+          });
+        // figure out which ones to remove
+        productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+          .map(({ id }) => id);
+        }
+
+        // run both actions
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove } }),
+          ProductTag.bulkCreate(newProductTags),
+        ]);
+      })
+      .then((updatedProductTags) => res.json(updatedProductTags))
+      .catch((err) => {
+        console.log(err);
+        res.status(400).json(err);
+      });
 });
 
 router.delete('/:id', (req, res) => {
